Removed the atomic integer since you look the id up in the HashMap and get the next empty id.
Removed the syncronized block in the Connection cleanup since the Map itself is syncronized and a iterator is a Snapshot of the Values in the Map. Since the iterator is the only remover in the Map thats ok.

Added packet 2 for sending Chunks over the Websocket.

Added packet 1 for requesting Chunks over the Websocket.

Changed the Spigot-APi to 1.7.2-R0.3-SNAPSHOT since R0.4 is not in the Repo